Experimental: introduce dot notation to call clojure functions
instead of having to define a new word for every clojure function
that's the same in kcats, we add a mechanism to invoke them
directly. It's analogous to clojure's mechanism to call java methods
directly.

Add words invoke1 and invoke2 to invoke the ToS word on the 1 or 2
args below it. The word should be preceded with a dot (like .count) -
the purpose here is to prevent squatting (malicious code defines a
word `count` and it clobbers the clojure invocation of `count` that
other code relies on). We also update `inscribe` to disallow adding
words that start with a dot.