Part way through a pipes implementation that is proving difficult
The problem is I'm trying to avoid putting stateful items on the
stack, and it's pretty hard to be able to compose using the language
itself, going straight from a value to create a pipe in any
circumstance.

I had tried to make everything functional at first (no stateful
objects on the stack), and then made an exception for pipes, but now
I'm wondering whether this is really a requirement.

The nice thing about pure values is it's portable - you could package
up a whole environment and ship it to another machine and continue
where it left off. The minute you include state you can no longer do
that (things like atomic values that might be shared with another env,
or a socket or a file, or anything that points to some real-world
location).

However I don't see how having stateful objects on the stack while
constructing a pipe is necessarily a problem - the pipe is already
non-portable. For example a socket connected to "localhost:1234" can't
be shipped off to another machine. First of all, "localhost" won't
mean the same thing to the other machine, second, even if it did, it
can't take over the old connection since it originated from somewhere
else.

Will try allowing state free-for-all, but will distinguish between
values and state in the spec.