Allow anything onto the stack
Makes creating pipes and dealing with potentially stateful
intermediate values a lot easier.

I think the main reason for values-only is for portability but as soon
as we involve pipes that goes out the door anyway.

Possible exception is pipes used for things like lazy sequences. Maybe
there will be a serialization protocol for stuff like that in the
future.