Working implementation of 'map'
It's a bit more involved than I expected.

The way it operates: snapshots the stack and clones it for every item
in the list. Runs the program on each item and packs the result into a
list.

There are other ways to do it, Thun appears to expand the program/item
pairs into the expression and evaluate them all. That seems kind of a
waste of memory for large lists. This implementation is probably
pretty inefficient in terms of CPU. Not parallelizable
either. Entirely possible it'll be replaced with a new impl at some
point, but at least there's one working impl.