Change the semantics of 'lookup' to take a keylist
Instead of a single key. This avoids the problem of having to unwrap a
single item list in order to look up values based on word-keys. Now
you can provide a list of keys to drill into like clojure's get-in.

TODO: step doesn't account for when the key is missing somewhere along
the way. It should stop right there as soon as whatever is given is
not an alist, and return nothing.