Change semantics of 'inscribe'
Now takes a k and v to assoc into the dictionary. The key is a bare
word, the value is an alist that should contain a 'definition' key
that is the program to substitute for that word during execution.

Can also have a spec key that is converted to clojure spec as follows:

[[s string] [i integer]]
would be like
(stack-spec :s ::string, :i ::integer)