Fix 'any?' example to match actual behavior
I'm not sure if this is the best fix - can either always return a
boolean, or return whatever truthy value comes from the underlying
clojure `or`. It's definitely more useful to give the actual value
instead of coercing to boolean but this may be surprising behavior to
less experienced users.