Change spec for the lexicon to a more natural syntax
spec is now a list of two items, input and output:

* What was on the top of the stack before (that the word
needs/uses). Top of stack is first.

* What is on top of the stack after (that the word changed). Top of
stack is first.

Each of those lists has items in the following format:

One of:

* A list of two items: the stack item's name, type. The name is only
used if it helps describe the item beyond its type, and/or the item is
moved in the stack and needs to be represented in both the input and
output. For example, in the word swap.

* A word (the type of item), when the name is not needed.

* An asterisk, meaning, 0 or more new stack items, that may replace 0
or more previous stack items. Use this when the word has arbitrary
stack effects. For example, 'dip' can change the stack arbitrarily but
always puts a particular item on top. So the output is [item *]