Use VecDeque everywhere, including env stack/expression
reason1 - evert'ing the stack with a list makes it a lot easier if
they're the same type, so we can just swap the memory locations. This
is important because evert is a *very* commonly used primitive and
copying is just not acceptable here.

reason2 - the environment itself is supposed to be a data structure
just like anything else in kcats, will be easier this way to nest envs
etc.