Start doing validation before axiom fn call, err msg improvements
Will use specs to do validation before doing function calls. Then
there's only one validation function.

Problem right now is that the lexicon contains the specs for the
builtin words, and we don't even read the lexicon until it's too late
to change the builtin objects to contain the specs. So the validation
is only working on derived words, not axiom words.

A way to solve this might be:

* Separate the builtin words' docs into a separate file eg builtins.kcats

* Read builtins.kcats when building the standard dictionary, similar
to how word lexicon does it

* Add another field to builtins, the parsed spec

* Call the validation as with other words, since the spec will be part
of the object already.