Remove the now-extraneous type and count checks
These are now done before calling any word functions, based on the
spec. So by the time the actual axiom function is called, we know the
arguments are good and don't have to test their presence and type.

Also stop using env_effect so much, it's simpler to use f_stack.