Not started: make errors first class values
This arose because of nested envs. Let's say we eval_step a nested
env, and it errors. what should the eval_step word put on the stack?

In our top level main(), we just print out the error but it's not
actually a *value*. I think it needs to be - I think we just put an
error on the stack and print it.

The question is, what happens after that? I think eval (the rust
fn) should just consider the program finished even if there's still
items in the expression. It'll print the stack and the error object
will be there. I think we should not consume any arguments for
whatever word caused the error, leave it in the expression and leave
the arguments. Then errors don't even need to report their context,
the error will already be shown in the context of where it occurred.

This also suggests a way to possibly recover from errors: the
environment executing the inner env, if it bails with an error, can
inspect the inner env, pop the error off the stack, fix whatever is
wrong on the stack, and proceed. For example, you could watch for
division_by_zero on the stack, and just clear the numbers and / word,
and put 0 on the stack. What's missing here is a way to specify
context though - if you wanted that error recovery only during
execution of a particular word, for example. And due to the way the
runtime works now, we don't know what words we're executing right
now (derived words have already been replaced with their definitions
in the expression).

Delegating to the outer env may be kind of limited - maybe we can have
ways of recovering from an error within the current env. For example,
in an 'if' expression, maybe in some cases, we want to just have the
condition end up as false if any errors occur.

Maybe a word 'recover' that executes a program, and if it errors, runs
another program. The problem there is, when an error does occur and
the "recovery" program is started, it's not really going to know what
to expect on the stack beneath the error. It would depend on where the
error occurred. In a procedural language, the execution stack is
unwound to where the error is caught. There has to be some kind of
guarantee here (perhaps using 'shield' type of stack-protection). One
possibility is to inject the program into a snapshot. If there's an
error on top afterward, grab it out of the snapshot and place it on
the stack, drop the snapshot, then call the recovery. The error would
have to keep the context of what went wrong. If the program succeeds,
evert and discard top item (dropping the saved original state). One
challege here is how to know when to stop execution (eg there is no
recover or it fails too). .The most perf hit I'm willing to take is to
check the rust item type on every step of invocation to see if it's an
error. But then how to further decide if we're recovering and we
shouldn't stop yet? It's possible to implement this by setting extra
state on the environment, but I'd rather not since it adds complexity.