Start using Rc for Lists
Instead of using actual lists, keep a reference to the list. When
modifying a list, use clone-on-write (if there are other references to
the list, clone it first). This should save a lot of copying during
loops and operations like 'snapshot' (which without optimization would
deep clone the entire stack, when usually only a few top items would
ever be modified).