Experimental: remove boolean values
This may be a bad idea, but I'm trying it out.

The rationale here is, do we really need booleans?

We have the concept of truthy/falsey without needing the actual values
true/false. We have the concept of "nothing" (represented as an empty
list), which is falsey. Without 'false', then everything else is
truthy. That seems a little less complicated than trying to support
booleans.

The part that is less elegant is what do we return from predicate
functions? Functionally speaking we could return any truthy value, but
we do have to pick something. 'true' was nice because it's a truthy
value that is easy to use by convention. We can return the *word*
true, and it's still by convention.

But then there's still the matter of whether we should treat "true"
special and define it as a self-inserting axiom word (just like an
actual boolean would have been).

And there's also the matter of [] and true looking rather different.