Make assoc get return owned value instead of reference
The reason for this is we have to return a reference to an Item, and
often we don't have an Item (eg if the actual object is an Error or
Environment - the fields are not Item enums). We can't just create an
Item and return a reference to it because it no longer has an owner
after that. We'd have to stash it somewhere it would continue to live
and we don't have anywhere.