Fixing some conversions to Entry
It occurs to me that requiring the word field to be on the Entry
object is not good. We do this to avoid looking up every word in a
hashmap at every step in execution.

There was an optimization where we lookup the Entry for a Word at read
time, so we can just execute it without a lookup. However now that
we're messing with the dictionary, this is no longer correct behavior
- the Entry could have been changed via 'lingo' between when it was
read in and executed.

I think will have to go back to hashmap lookups and try to find other
ways to optimize.