First cut at a interactive repl mode
The binary will take length-prefixed inputs, and print length prefixed
outputs. It's designed to be used with an IDE client and a first cut
at an elisp client is included. Both untested so far.

Another mode might be needed to go clientless (just a terminal) where
inputs are separated by \n or something.