Take returns Option<Item> - moving towards removing closed?
I think the rust idiom of only knowing if a pipe is closed by trying
to operate on it and handling the error is correct to adopt in kcats -
avoids race conditions where the pipe state changes between when you
check if it's closed and when you try to put/take.

Eventually we want 'step' to iterate over pipes and it continues as
long as the 'next' returns something. When the pipe is closed, it'll
return None, and stop the iteration. 'assemble' will then be written
in terms of 'step', and then the word 'closed?' will be unused and we
can remove it.

We will alias 'close' to 'discard' and document closing pipes by
example there.