Express map and filter in terms of transducers
This has the effect of disallowing multi-arity functions for mapping
or filtering. To have the same effect you need to capture the values
within the function itself (closing over) and have the list item
placed on top. The function can no longer expect more than one item on
the stack when it is called.