Store executing programs in stacked format, with less mutation
Note, there are some problems making this very efficient:

* internal mutation of the program state - can't mutate a vec element
in an Arc. Have to pop off the top item, mutate to increment the
counter, and push back.

* Can't directly read the next instruction. We have to clone it. Tried
using unsafe etc, but it can't be done. Would have to infect the whole
codebase with lifetime params and would be a huge mess.