Modify generate to support self-cleaning generators
When generate sees that the generator produced 'nothing', it will drop
the remaining copy of the program. 'nothing' is used to signal that no
more values will be produced from the generator so dropping it will
not only free up resources, but allows the generator to get out of the
way of access to lower items in the stack.

This works recursively so that layered generators should all clean
themselves up at the end-of-stream. No longer need to `shield` the
whole chain of generators.

However it does put an extra burden on creators of generators, to
'self-clean' - the generator's program should check if the value was
'nothing' and drop the state.

A decorator should be created to help with this, it should be able to
wrap any generator, and drop the state if it produced 'nothing'. May
need a different one for each arity (count of state items). Or maybe
just one that takes a count arg.