Separate word for autoformatting code
If we always do this on `emit` then we lose the determinism we need
for cryptographic stuff. The crypto functions should be able to get
serialized data as a single line, so that minor changes to the
autoformatting rules don't break cryptographic hashes and other
guarantees.

We should be able to rely on plain edn emit being consistent across
time, hopefully.