Fix bug where current instruction where error occurred is not saved
The issue is the current instruction is popped off the program, before
we know whether an error will occur. If an error does occur, we have
no convenient way to report what word was executing at the time.

The solution is for eval_step not to pop the word off, the individual
axiom functions will do it after they're sure the result is an Ok
value. If it's an Err, they leave the word on the program so that it
can be reported and potentially, retried later.