Finished first impl of template
This is the equivalent of a macro in lisp.

The ToS is the form you'd normally syntax-quote in lisp. Under that is
the values you intend to be able to unquote (as an association). The
unquote forms are [poke k] and [splice k] where k is a key in the
values assoc. It looks up the value v for key k. If you use poke, it
replaces the unquote form with v, and if you use splice, it splices in
v (and this assumes v is a list, will error if it isn't).