`decide` no longer depend on execution derived words
We want to load `decide` early since a lot of stuff uses it (like
'encode') so we need for it to not depend on anything loaded late,
like 'if' or 'dipdown'. So we replace those words with their
definitions inside the rust impl of `decide`. Also update the
definitions of dipdown, dipdeep to not depend on `decorated`.