Fix issue where Arc is always cloned even when there's only one copy.
Usually we intend to consume the contents when we access the inner(),
so we actually don't need that method. We can just borrow the
contents, and if we need owned values of the contents, we explicitly
clone them (usually using iter().cloned() or some such, for lists and
other iterables).

The fix didn't seem to make much of a performance difference that I
could see, though.