WIP fix partition
One issue is that it relies on other generators in its own module and
while we are loading modules one at a time, during the actual
execution of the module, we don't have that module available.

So the solution should be to break up the generators module into core
and then other generators that depend on the core ones.