During unwind, append to error's unwound field instead of overwrite
The problem is when we're executing in an inner env, and it gets an
unhandled error. We then promote that value to the top level
stack. The runtime sees there's an unhandled error and begins to
unwind the program but instead of adding to whatever's already in the
error, it overwrites the value, losing the context of where things
went wrong to produce the error.

However this is still not ideal because the idea of saving the
'unwound' field is that you can re-execute the program
later (presumably after the problem has been resolved). With this fix,
we end up with a frankenstein program that was produced by different
environments and possibly cannot be rerun anymore - the dictionary may
not match the original anymore.

What might be better is a way of unwinding while keeping the nested
structure. However this may be complex to deal with.