Update let comments, add drain, extractor, evaluator
drain: drains a generator (for side effects), keeping nothing

extractor: takes a generator stack wrapped in a list and converts it
to a single normal generator. (use if the stack is deep and the top
needs to access something deep in the stack)

evaluator: takes a generator of program snippets, and an
environment. Returns a generator of environments representing the
outcome of evaluating those program snippets.