Fix design error where converting str to Word couldn't fail
Not all str's are valid words. When we're converting a compile-time
literal, we allow it to panic. If it's a runtime value we surface the
parse error, and we always use the edn parser to determine if it's a
valid Word (symbol). The parser is still probably more lenient than I
would prefer, but that can be fixed separately.