Fix until: body shouldn't receive pred result, compile properly
two fixes here:

1) because the pred logic is reversed (only stop when pred returns
truthy) the body of the loop will always receive []. So we can just
pre-emptively drop it so the passed body program never sees it.

2) We were executing the pre-loop body once separately from the while
loop execution (one at 'compile time' and later at 'runtime' when the
while loop is executed). They should be part of the same program, and
now they are.