Attempt to use build.rs to lock axiom words to namespace
I don't think this is needed. Instead: make all modules functions of
the dictionary. Enforce namespacing at inscribe time, then when
merging, do the checking as to whether we're merging a bare function
with its counterpart.

We do still need to properly namespace the builtins so we do need to
save the hash and filename of each, so we can refer to the filename
later.